canary/requirements.txt
viam-sdk
pymongo